MFC part of rev. 1.38: correct type for the 3rd argument to sysctlbyname().

+++ usr.bin/top/machine.c
@@ -226,7 +226,7 @@ struct statics *statics;
 {
     register int i = 0;
     register int pagesize;
-    int modelen;
+    size_t modelen;
     struct passwd *pw;
 
     modelen = sizeof(smpmode);
